Tests, and cleanup
