+++ Dockerfile
@@ -15,6 +15,7 @@ ENV \
     DOTNET_CLI_TELEMETRY_OPTOUT=true \
     HASS_RUN_PROJECT_FOLDER=/usr/src/Service \
     HASS_HOST=localhost \
+    HASSCLIENT_MSGLOGLEVEL=Default \
     HASS_PORT=8123 \
     HASS_TOKEN=NOT_SET \
     HASS_DAEMONAPPFOLDER=/data
