+++ src/App/NetDaemon.App/Common/Reactive/ObservableExtensionMethods.cs
@@ -1,0 +1,22 @@
+﻿using System;
+using System.Reactive.Linq;
+
+namespace JoySoftware.HomeAssistant.NetDaemon.Common.Reactive
+{
+    /// <summary>
+    ///     Extension methods for Observables
+    /// </summary>
+    public static class ObservableExtensionMethods
+    {
+        /// <summary>
+        ///     Is same for timespan time
+        /// </summary>
+        /// <param name="observable"></param>
+        /// <param name="span"></param>
+        /// <returns></returns>
+        public static IObservable<(EntityState Old, EntityState New)> NDSameStateFor(this IObservable<(EntityState Old, EntityState New)> observable, TimeSpan span)
+        {
+            return observable.Throttle(span);
+        }
+    }
+}
