+++ src/Daemon/NetDaemon.Daemon/Daemon/CodeManager.cs
@@ -5,10 +5,7 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
-using System.Reflection;
 using System.Runtime.CompilerServices;
-using System.Runtime.Loader;
-using System.Threading.Tasks;
 
 [assembly: InternalsVisibleTo("NetDaemon.Daemon.Tests")]
 
@@ -18,9 +15,8 @@ namespace JoySoftware.HomeAssistant.NetDaemon.Daemon
     {
         private readonly string _codeFolder;
         private readonly ILogger _logger;
-        private IEnumerable<Type>? _loadedDaemonApps;
-
         private readonly YamlConfig _yamlConfig;
+        private IEnumerable<Type>? _loadedDaemonApps;
 
         /// <summary>
         ///     Constructor
@@ -69,6 +65,5 @@ namespace JoySoftware.HomeAssistant.NetDaemon.Daemon
             }
             return result;
         }
-
     }
 }
