+++ src/Daemon/NetDaemon.Daemon/Daemon/INetDaemonHost.cs
@@ -1,6 +1,5 @@
 using JoySoftware.HomeAssistant.NetDaemon.Common;
 using System;
-using System.Collections.Generic;
 using System.Runtime.CompilerServices;
 using System.Threading;
 using System.Threading.Tasks;
@@ -9,11 +8,30 @@ using System.Threading.Tasks;
 
 namespace JoySoftware.HomeAssistant.NetDaemon.Daemon
 {
+
     /// <summary>
     ///     The interface that interacts with the daemon host main logic
     /// </summary>
     public interface INetDaemonHost : INetDaemon
     {
+        /// <summary>
+        ///     Initializes the NetDaemon
+        /// </summary>
+        Task Initialize();
+
+        /// <summary>
+        ///     Listens to the given service in the netdaemon domain. Those subscritions
+        ///     are used internally and are not postponed during reloading service daemons.
+        /// </summary>
+        /// <param name="service"> The name of the service. </param>
+        /// <param name="action"> The async. action that is executed. </param>
+        void ListenCompanionServiceCall(string service, Func<dynamic?, Task> action);
+
+        /// <summary>
+        ///     Reload all apps
+        /// </summary>
+        Task ReloadAllApps();
+
         /// <summary>
         ///     Runs the netdaemon host asynchronous.
         /// </summary>
@@ -26,20 +44,6 @@ namespace JoySoftware.HomeAssistant.NetDaemon.Daemon
         Task Run(string host, short port, bool ssl, string token, CancellationToken cancellationToken);
 
         /// <summary>
-        ///     Stops the netdaemon host asynchronous.
-        /// </summary>
-        /// <returns> The operational task. </returns>
-        Task Stop();
-
-        /// <summary>
-        ///     Listens to the given service in the netdaemon domain. Those subscritions
-        ///     are used internally and are not postponed during reloading service daemons.
-        /// </summary>
-        /// <param name="service"> The name of the service. </param>
-        /// <param name="action"> The async. action that is executed. </param>
-        void ListenCompanionServiceCall(string service, Func<dynamic?, Task> action);
-
-        /// <summary>
         ///     Sends the status of the netdaemon host to Home Assistant.
         /// </summary>
         /// <param name="numberOfLoadedApps"> The number of loaded apps. </param>
@@ -48,26 +52,13 @@ namespace JoySoftware.HomeAssistant.NetDaemon.Daemon
         Task SetDaemonStateAsync(int numberOfLoadedApps, int numberOfRunningApps);
 
         /// <summary>
-        ///     Initializes the NetDaemon
+        ///     Stops the netdaemon host asynchronous.
         /// </summary>
-        Task Initialize();
-
+        /// <returns> The operational task. </returns>
+        Task Stop();
         /// <summary>
         ///     Clears all app instances registered
         /// </summary>
         Task UnloadAllApps();
-
-        /// <summary>
-        ///     Reload all apps
-        /// </summary>
-        Task ReloadAllApps();
     }
-
-    public interface IInstanceDaemonApp
-    {
-
-        IEnumerable<INetDaemonAppBase> InstanceDaemonApps();
-        int Count { get; }
-    }
-
 }
